🚀 RELEASE: Frontend basic version
