+++ .env.example
@@ -7,5 +7,7 @@ NEXT_PUBLIC_CONTRACT_ADDR=0x32888423d4B8Aa9a555C9DBAd82a898646B69925
 NEXT_PUBLIC_APP_ID=app_staging_27f3e5996cfe801b96d47df2ffa35053
 NEXT_PUBLIC_WALLETCONNECT_PROJECTID=3c14b563100c6a967d876ca158ee21d2
 
+NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID=e65337e14062fd6da6689403796190bb
+
 # Rename this file to `.env` and update the key with a funded account for contract deployments
 DEPLOYER_KEY='0xFFFFFFFFFFFFFFFFFFFFFFFFFFF0000000000000000000000000000000000000'
