+++ .prettierrc.json
@@ -5,5 +5,5 @@
     "printWidth": 150,
     "bracketSameLine": true,
     "useTabs": false,
-    "tabWidth": 2
+    "tabWidth": 1
 }
