+++ src/pages/createlink.tsx
@@ -1,0 +1,34 @@
+import { Box, Text, Heading } from '@chakra-ui/react';
+import type { NextPage } from 'next';
+import Head from 'next/head';
+
+const Home: NextPage = () => {
+  // TODO: Fetch proof from shared state
+  const proof = 'aaaa';
+
+  // Don't show page if there's no WorldId
+  // TODO: Link to download Worldcoin App
+  // TODO: QR Code for worldcoin app
+  if (!proof)
+    return (
+      <Box>
+        <Heading>You must log in with Eth Address and Worldcoin</Heading>
+        <Text>Connect here!</Text>
+      </Box>
+    );
+
+  return (
+    <Box maxW="container.md" mx="auto" p={8}>
+      <Head>
+        <title>Create Campaign</title>
+        <meta
+          content="Generated by @rainbow-me/create-rainbowkit"
+          name="description"
+        />
+        <link href="/favicon.ico" rel="icon" />
+      </Head>
+    </Box>
+  );
+};
+
+export default Home;
