+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
-import { Text } from '@chakra-ui/react'
-import { Head } from 'components/layout/Head'
-import { HeadingComponent } from 'components/layout/HeadingComponent'
-import { LinkComponent } from 'components/layout/LinkComponent'
+import { Text } from '@chakra-ui/react';
+import { Head } from 'components/layout/Head';
+import { HeadingComponent } from 'components/layout/HeadingComponent';
+import { LinkComponent } from 'components/layout/LinkComponent';
 
 export default function Home() {
   return (
@@ -12,10 +12,11 @@ export default function Home() {
         <HeadingComponent as="h2">Create a campaign</HeadingComponent>
         <Text>Quickly ship Web3 Apps ⚡</Text>
         <Text py={4}>
-          <LinkComponent href="examples">View examples</LinkComponent> to bootstrap development.
+          <LinkComponent href="examples">View examples</LinkComponent> to
+          bootstrap development.
         </Text>
         <HeadingComponent as="h2">Create ref link</HeadingComponent>
       </main>
     </>
-  )
+  );
 }
